[fix] 인터셉터에서 기업부분 오류 수정 (#39)
* 받음

* 공고 상세 수정

* ì유저가 같은회사에 지원 못하게 수정

* ì예외처리 이인터셉터

* 이인터셉터 예외처리 로그인만

* 기업예외처리완료

* 기업예외처리 추가

* [fix] BoardController id type

+++ src/main/java/com/metacoding/springrocketdanv1/board/BoardController.java
@@ -49,7 +49,7 @@ public class BoardController {
     }
 
     @PostMapping("/board/delete")
-    public String delete(@RequestParam("id") Long id) { // boardId를 가져와서 삭제
+    public String delete(@RequestParam("id") Integer id) { // boardId를 가져와서 삭제
         boardService.글삭제하기(id);
         return "redirect:/board";
     }
